third_party/WebKit/LayoutTests/storage/quota/storageinfo-missing-arguments.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>

description('Ensures that we get exceptions when we do not include required arguments for webkitStorageInfo methods.');

if (window.webkitStorageInfo) {
    shouldThrow('webkitStorageInfo.requestQuota()', '"TypeError: Failed to execute \'requestQuota\' on \'DeprecatedStorageInfo\': 2 arguments required, but only 0 present."');
    shouldThrow('webkitStorageInfo.requestQuota(webkitStorageInfo.TEMPORARY)', '"TypeError: Failed to execute \'requestQuota\' on \'DeprecatedStorageInfo\': 2 arguments required, but only 1 present."');
    shouldThrow('webkitStorageInfo.queryUsageAndQuota()', '"TypeError: Failed to execute \'queryUsageAndQuota\' on \'DeprecatedStorageInfo\': 1 argument required, but only 0 present."');
} else
    debug("This test requires window.webkitStorageInfo.");

</script>
