third_party/blink/web_tests/storage/quota/storageinfo-missing-arguments-expected.txt

Ensures that we get exceptions when we do not include required arguments for webkitStorageInfo methods.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS webkitStorageInfo.requestQuota() threw exception TypeError: Failed to execute 'requestQuota' on 'DeprecatedStorageInfo': 2 arguments required, but only 0 present..
PASS webkitStorageInfo.requestQuota(webkitStorageInfo.TEMPORARY) threw exception TypeError: Failed to execute 'requestQuota' on 'DeprecatedStorageInfo': 2 arguments required, but only 1 present..
PASS webkitStorageInfo.queryUsageAndQuota() threw exception TypeError: Failed to execute 'queryUsageAndQuota' on 'DeprecatedStorageInfo': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
